docker: fix python installation
This change fixes some issues with the Python installation within a
Docker container containing snapcraft built from the docker/Dockerfile
in the project as described here:

https://bugs.launchpad.net/snapcraft/+bug/1956216

Background: The image built from the Dockerfile used to work fine. Some
weeks/months ago, it stopped working out of a sudden. Digging into the
failures, it seems the Python installation that comes with recent
snapcraft versions seems to be somehow broken. This change remedies this
by fixing up the Python installation after installation within the
Docker image. For this, we:

- Adjust PATH so the correct Python interpreter will be used.
- Fix a symlink.
- Add a *.pth file to make sure the Python 3 in the snapcraft snap finds
  its own site-packages (for some reason, the site-packages is not in
  sys.path by default).

LP: #1956216